+++ views/public/join.php
@@ -8,7 +8,6 @@ use yii\helpers\Url;
 /* @var $session \k7zz\humhub\bbb\models\Session|null */
 /* @var $token string */
 /* @var $msg string|null */
-/* @var $reload bool|null */
 
 $bundle = BBBAssets::register($this);
 $this->setPageTitle(($session && $session->title) ? $session->title : Yii::t('BbbModule.base', 'Join session'));
@@ -17,7 +16,9 @@ $imageUrl = ($session && $session->image_file_id)
     ? Url::to(['/bbb/public/download', 'id' => $session->id, 'type' => 'image', 'inline' => true])
     : $bundle->baseUrl . '/images/conference.png';
 ?>
-<div id="layout-content">
+<div id="layout-content"
+    data-bbb-check-running="<?= Html::encode(Url::to(['/bbb/public/is-running', 'token' => $token])) ?>"
+    data-bbb-redirect-on-running>
     <div class="container-fluid">
         <div class="row">
             <div class="col-md-8 offset-md-2 col-lg-6 offset-lg-3">
@@ -44,44 +45,21 @@ $imageUrl = ($session && $session->image_file_id)
 
                         <?php if ($msg): ?>
                             <div class="alert alert-danger"><?= Html::encode($msg) ?></div>
-                            <?php if ($reload): ?>
-                                <div id="bbb-public-waiting" class="alert alert-info">
-                                    <i class="fa fa-spinner fa-spin"></i>
-                                    <?= Yii::t('BbbModule.base', 'You will be redirected when the session starts.') ?>
+                            <div class="bbb-waiting" class="alert alert-info"
+                                style="display: <?= $running ? 'none' : '' ?>;">
+                                <i class="fa fa-spinner fa-spin"></i>
+                                <?= Yii::t('BbbModule.base', 'You will be redirected when the session starts.') ?>
+                            </div>
+                            <div class="bbb-running" style="display: <?= $running ? '' : 'none' ?>;">
+                                <div class="alert alert-success">
+                                    <i class="fa fa-check"></i>
+                                    <?= Yii::t('BbbModule.base', 'The session has started!') ?>
                                 </div>
-                                <div id="bbb-public-join-ready" style="display:none;">
-                                    <div class="alert alert-success">
-                                        <i class="fa fa-check"></i>
-                                        <?= Yii::t('BbbModule.base', 'The session has started!') ?>
-                                    </div>
-                                    <a href="#" class="btn btn-success btn-lg w-100" id="bbb-public-join-btn">
-                                        <i class="fa fa-sign-in"></i>
-                                        <?= Yii::t('BbbModule.base', 'Join now') ?>
-                                    </a>
-                                </div>
-                                <script <?= Html::nonce() ?>>
-                                    (function poll() {
-                                        setTimeout(function() {
-                                            fetch(<?= json_encode(Url::to(['/bbb/public/is-running', 'token' => $token])) ?>)
-                                                .then(function(r) { return r.json(); })
-                                                .then(function(data) {
-                                                    if (data.running) {
-                                                        document.getElementById('bbb-public-waiting').style.display = 'none';
-                                                        document.getElementById('bbb-public-join-ready').style.display = '';
-                                                    } else {
-                                                        poll();
-                                                    }
-                                                })
-                                                .catch(function() { poll(); });
-                                        }, 5000);
-                                    })();
-
-                                    document.getElementById('bbb-public-join-btn').addEventListener('click', function(e) {
-                                        e.preventDefault();
-                                        window.open(window.location.href, '_blank');
-                                    });
-                                </script>
-                            <?php endif; ?>
+                                <a href="#" class="btn btn-success btn-lg w-100" id="bbb-public-join-btn">
+                                    <i class="fa fa-sign-in"></i>
+                                    <?= Yii::t('BbbModule.base', 'Join now') ?>
+                                </a>
+                            </div>
                         <?php else: ?>
                             <p><?= Yii::t('BbbModule.base', 'Please enter your name below:') ?></p>
                             <form method="get" action="<?= Html::encode($action) ?>">